+++ python/instrumentation/openinference-instrumentation-anthropic/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.1.3](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-anthropic-v2.1.2...python-openinference-instrumentation-anthropic-v2.1.3) (2026-09-09)
+
+
+### Bug Fixes
+
+* bump openinference-instrumentation minimum to &gt;=0.1.61 ([#3680](https://github.com/Arize-ai/openinference/issues/3680)) ([ba0fc57](https://github.com/Arize-ai/openinference/commit/ba0fc57cc735c8d4d85c125403b781dd6c998ad6))
+
 ## [2.1.2](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-anthropic-v2.1.1...python-openinference-instrumentation-anthropic-v2.1.2) (2026-09-04)